+++ src/client/dictionaries/index.js
@@ -1,0 +1,15 @@
+const {YandexDictionary} = require('./YandexDictionary')
+const {LocalDictionary} = require('./LocalDictionary')
+
+function getDictionary() {
+    return new YandexDictionary()
+}
+
+function getLocalDictionary() {
+    return new LocalDictionary()
+}
+
+module.exports = {
+    getDictionary,
+    getLocalDictionary,
+}