+++ src/config/index.js
@@ -1,0 +1,10 @@
+const {SOURCE, TARGET} = require('./language')
+const SERVER_PORT = 3000
+
+module.exports = {
+    LANGUAGE: {
+        SOURCE,
+        TARGET,
+    },
+    SERVER_PORT,
+}