+++ src/config/language.js
@@ -1,0 +1,16 @@
+const SOURCE = {
+    INPUT_SELECTOR: '#sourceText',
+    LANGUAGE_CODE: 'de',
+    LANGUAGE_FULL: 'Deutsch',
+}
+
+const TARGET = {
+    INPUT_SELECTOR: '#targetText',
+    LANGUAGE_CODE: 'en',
+    LANGUAGE_FULL: 'English',
+}
+
+module.exports = {
+    SOURCE,
+    TARGET,
+}