+++ src/service/dictionaries/index.js
@@ -1,9 +1,0 @@
-const {YandexDictionary} = require('./YandexDictionary')
-
-function getDictionary() {
-    return new YandexDictionary()
-}
-
-module.exports = {
-    getDictionary
-}